refactor: merge store
